feat: Add reservation screen to navbar
The code changes include importing the `ReservationScreen` and adding it to the list of pages in the `NavbarScreen` widget. The `AppBar` title is also updated to display "Reservasi Online" when the `ReservationScreen` is selected. Additionally, the background color of the `BottomNavigationBar` is set to white.

Based on the recent repository commits, it seems that there have been updates to assets and colors, as well as the addition of hero images. However, these changes are not directly related to the code changes in this commit.

Please review and make any necessary adjustments before committing.