feat: Add history screen to navbar
The code changes include importing the `HistoryScreen` and adding it to the list of pages in the `NavbarScreen` widget. The `AppBar` title is also updated to display "History" when the `HistoryScreen` is selected.

Based on the recent repository commits, it seems that there have been updates to assets and colors, as well as the addition of hero images. However, these changes are not directly related to the code changes in this commit.