feat: Set initial selected index based on user role
This commit modifies the `_NavbarScreenState` class in `navbar_screen.dart` to set the initial selected index of the bottom navigation bar based on the user's role. If the role is 'admin' or 'staff', the default page is set to index 3. If the role is 'doctor', the default page is also set to index 3. For all other roles, the default page is set to index 0.