feat: Update home screen to limit the number of articles displayed
This commit modifies the `HomeScreenState` class in the `home_screen.dart` file. It updates the `itemCount` property of the `ListView.separated` widget to limit the number of articles displayed to a maximum of 5. This change improves the performance of the home screen by reducing the number of items rendered in the list.

Please review and make any necessary adjustments before committing.