feat: Add back button to article screen
This commit adds a back button to the article screen in the `_ArticleScreenState` class. The back button is represented by an arrow icon and allows the user to navigate back to the previous screen. This enhances the user experience by providing a convenient way to go back without having to use the device's back button.