feat: Add action icon to assets and update master screen
This commit adds a new action icon to the assets in the `assets/icons` directory. The icon file is named `action.svg`. Additionally, the master screen in the `features/master/presentation/pages/master_screen.dart` file is updated to use the `DataReservationScreen` when the "Reservasi Online" button is pressed, instead of the previous `DataServiceScreen`.