feat: Update searchChanged method in DataReservationScreen
This commit updates the `searchChanged` method in the `_DataReservationScreenState` class of the `data_reservation_screen.dart` file. It now checks if the `value` is not empty and triggers the `getReservationDataByName` event in the `DataReservationBloc` with the `searchController.text` as the parameter. If the `value` is empty, it triggers the `getReservationData` event. This change allows the app to fetch reservation data based on the search input and improves the user experience.