feat: Add functionality to add patients from navbar screen
This commit adds functionality to the navbar screen to allow users to add patients directly from the navbar. When the user is on the home or patients screen, they will see an "Add" button in the app bar. Clicking on this button will navigate them to the add patient screen. This change improves the user experience by providing a convenient way to add patients without navigating through multiple screens.