feat: Add API endpoints for adding and editing doctors
This commit adds two new API endpoints, `addDoctor` and `editDoctor`, to the `MasterRemoteDatasources` class in the `master_remote_datasources.dart` file. These endpoints allow the app to add and edit doctor data in the backend server. This change enhances the functionality of the app by providing the ability to perform CRUD operations on doctor data.