feat: Add deleteDoctor event to AddAndEditDoctorBloc
This commit adds the deleteDoctor event to the AddAndEditDoctorBloc in the add_and_edit_doctor_bloc.dart file. The event allows the app to delete a doctor from the backend server using the doctor's ID. This change enhances the functionality of the app by providing the ability to perform the delete operation on doctor data.