feat: Add liquid_pull_refresh package for pull-to-refresh functionality
This commit adds the `liquid_pull_refresh` package (`^1.0.6`) as a new dependency in the `pubspec.yaml` file. The package provides pull-to-refresh functionality, allowing users to refresh the content of the app by pulling down on the screen. This change enhances the user experience by providing a convenient and intuitive way to update the displayed data.