feat: Add historyImage field to MasterReservation model
This commit adds the `historyImage` field to the `MasterReservation` model in the `master_reservation_response_model.dart` file. The `historyImage` field represents the image associated with the reservation history. This change enhances the functionality of the app by providing the ability to store and retrieve the history image for a reservation.