chore: Update dependencies and add lazy loading for images
Update dependencies in pubspec.yaml to the latest stable versions of build_runner, freezed, and flutter_gen_runner. This ensures that the app uses the most up-to-date packages for development. Additionally, add lazy loading for images in the add_and_edit_doctor_bloc file to improve page load performance and optimize resource usage.